requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fastapi==0.118.2
fonttools==4.60.1
h11==0.16.0
idna==3.10
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.7
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
regex==2025.10.23
requests==2.32.5
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
websocket-client==1.9.0
